requirements/dev.txt
black
isort
django-extensions
model-bakery
pytest-django